+++ LibiadaMusic/Characteristics/WithinWordRemoteness.cs
@@ -21,17 +21,17 @@ namespace LibiadaMusic.Characteristics
 
             foreach (Fmotiv fmotiv in FmChain.FmotivList)
             {
-                Chain notechain = new Chain(fmotiv.PauseTreatment(ParamPauseTreatment.Ignore).TieGathered().NoteList.Count);
-                for (int i = 0; i < fmotiv.PauseTreatment(ParamPauseTreatment.Ignore).TieGathered().NoteList.Count; i++)
+                Chain notechain = new Chain(fmotiv.PauseTreatment((int)ParamPauseTreatment.Ignore).TieGathered().NoteList.Count);
+                for (int i = 0; i < fmotiv.PauseTreatment((int)ParamPauseTreatment.Ignore).TieGathered().NoteList.Count; i++)
                 {
                     temp = "";
-                    foreach (Pitch pitch in fmotiv.PauseTreatment(ParamPauseTreatment.Ignore).TieGathered().NoteList[i].Pitch)
+                    foreach (Pitch pitch in fmotiv.PauseTreatment((int)ParamPauseTreatment.Ignore).TieGathered().NoteList[i].Pitch)
                     {
                         temp += Convert.ToString(pitch.Midinumber);
                     }
                     //TODO: переделать нормально чтоб цепочка складывалась из ValueNote, а не как попало
                     notechain[i] = new ValueString(temp + " " +
-                        Convert.ToString(fmotiv.PauseTreatment(ParamPauseTreatment.Ignore).TieGathered().NoteList[i].Duration.Value * 10000000));
+                        Convert.ToString(fmotiv.PauseTreatment((int)ParamPauseTreatment.Ignore).TieGathered().NoteList[i].Duration.Value * 10000000));
                 }
 
                 Characteristic R = new Characteristic(new Depth());
