+++ LibiadaMusicTest/BorodaDivider/FmotivIdentificatorTest.cs
@@ -48,12 +48,12 @@ namespace LibiadaMusicTest.BorodaDivider
 
         FmotivIdentificator fmid = new FmotivIdentificator();
 //        FmotivChain fmidentedchain = new FmotivChain();
-        Assert.AreEqual(0, fmid.GetIdentification(fmchain1,ParamPauseTreatment.Ignore,ParamEqualFM.Sequent).FmotivList[0].Id);
-        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[1].Id);
-        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[2].Id);
-        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[3].Id);
-        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[4].Id);
-        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[5].Id);
+        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[0].Id);
+        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[1].Id);
+        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[2].Id);
+        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[3].Id);
+        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[4].Id);
+        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[5].Id);
 
     }
     [TestMethod]
@@ -98,12 +98,12 @@ namespace LibiadaMusicTest.BorodaDivider
 
         FmotivIdentificator fmid = new FmotivIdentificator();
         //        FmotivChain fmidentedchain = new FmotivChain();
-        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[0].Id);
-        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[1].Id);
-        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[2].Id);
-        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[3].Id);
-        Assert.AreEqual(2, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent).FmotivList[4].Id);
-        Assert.AreEqual(3, fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent ).FmotivList[5].Id);
+        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[0].Id);
+        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[1].Id);
+        Assert.AreEqual(0, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[2].Id);
+        Assert.AreEqual(1, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[3].Id);
+        Assert.AreEqual(2, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[4].Id);
+        Assert.AreEqual(3, fmid.GetIdentification(fmchain1, (int)ParamPauseTreatment.Ignore, (int)ParamEqualFM.Sequent).FmotivList[5].Id);
 
     }
     }
