+++ LibiadaMusicTest/ScoreModel/TieTest.cs
@@ -9,22 +9,22 @@ namespace LibiadaMusicTest.ScoreModel
         [TestMethod]
         public void TestTie1()
         {
-            Assert.AreEqual(-1, Tie.None);
+            Assert.AreEqual(-1, (int)Tie.None);
         }
         [TestMethod]
         public void TestTie2()
         {
-            Assert.AreEqual(0, Tie.Start);
+            Assert.AreEqual(0, (int)Tie.Start);
         }
         [TestMethod]
         public void TestTie3()
         {
-            Assert.AreEqual(1, Tie.Stop);
+            Assert.AreEqual(1, (int)Tie.Stop);
         }
         [TestMethod]
         public void TestTie4()
         {
-            Assert.AreEqual(2, Tie.StartStop);
+            Assert.AreEqual(2, (int)Tie.StartStop);
         }
     }
 }
